requirements.txt
flask
torch
torchvision
Pillow
numpy
scikit-learn
matplotlib
seaborn
gunicorn
